# format any type