```console
$ benchstat a.txt b.txt                             
goos: darwin
goarch: arm64
pkg: go.expect.digital/mf2/template
cpu: Apple M2 Max
                   │    a.txt    │               b.txt                │
                   │   sec/op    │   sec/op     vs base               │
Template_Sprint-12   2.703µ ± 0%   2.646µ ± 1%  -2.09% (p=0.000 n=10)

                   │    a.txt     │                b.txt                │
                   │     B/op     │     B/op      vs base               │
Template_Sprint-12   4.588Ki ± 0%   4.578Ki ± 0%  -0.21% (p=0.000 n=10)

                   │   a.txt    │               b.txt               │
                   │ allocs/op  │ allocs/op   vs base               │
Template_Sprint-12   93.00 ± 0%   92.00 ± 0%  -1.08% (p=0.000 n=10)
```